实现大部分功能

+++ src/main/java/xin/shaonianyou/forum/entity/Post.java
@@ -4,23 +4,23 @@ import java.io.Serializable;
 import java.util.Date;
 
 public class Post implements Serializable {
-    private int id;             //主键
+    private long id;             //主键
     private String title;       //帖子标题
     private String body;        //帖子内容
     private Date date_created;  //发帖日期
     private long user_id;       //发帖人
-    private  int module_id;     //模块ID
-    private int category_id;    //分类ID
-    private int comment_count;  //回复统计
-    private int hit_count;      //点赞统计
+    private long module_id;     //模块ID
+    private long category_id;    //分类ID
+    private long comment_count;  //回复统计
+    private long hit_count;      //点赞统计
 
     private static final long serialVersionUID = 1L;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -56,35 +56,35 @@ public class Post implements Serializable {
         this.user_id = user_id;
     }
 
-    public int getModule_id() {
+    public long getModule_id() {
         return module_id;
     }
 
-    public void setModule_id(int module_id) {
+    public void setModule_id(long module_id) {
         this.module_id = module_id;
     }
 
-    public int getCategory_id() {
+    public long getCategory_id() {
         return category_id;
     }
 
-    public void setCategory_id(int category_id) {
+    public void setCategory_id(long category_id) {
         this.category_id = category_id;
     }
 
-    public int getComment_count() {
+    public long getComment_count() {
         return comment_count;
     }
 
-    public void setComment_count(int comment_count) {
+    public void setComment_count(long comment_count) {
         this.comment_count = comment_count;
     }
 
-    public int getHit_count() {
+    public long getHit_count() {
         return hit_count;
     }
 
-    public void setHit_count(int hit_count) {
+    public void setHit_count(long hit_count) {
         this.hit_count = hit_count;
     }
